fix: call github API multiple times
The collect info task that fetches the runners from the Github API had
the `run_once` option. This meant that a single API call was made and
response reused when provisioning all hosts in an inventory.

This can break when configuring an inventory of repository and
organisation level runners. An API payload with a list of organisation
runners might be used when configuring a repository level runner.